+++ Tests/New-YTIssue.Tests.ps1
@@ -34,24 +34,30 @@ BeforeAll {
     {
         [CmdletBinding()]
         param(
-            [String] $Summary,
-            [String] $Description,
-            [String] $Project,
-            [String] $HasParent
+            [String] $HasSummary,
+            [String] $HasDescription,
+            [String] $InProject,
+            [String] $HasParent,
+            [hashtable] $HasCustomFields
         )
 
         $script:result.idReadable | Should -Not -BeNullOrEmpty
 
-        if ($Summary)
+        if ($HasSummary)
         {
-            $script:result.summary | Should -Be $Summary
+            $script:result.summary | Should -Be $HasSummary
         }
 
         $script:result.project.id | Should -Be $script:project.id
 
-        if ($Description)
+        if ($InProject)
         {
-            $script:result.description | Should -Be $Description
+            $script:result.project.shortName | Should -Be $InProject
+        }
+
+        if ($HasDescription)
+        {
+            $script:result.description | Should -Be $HasDescription
         }
 
         if ($HasParent)
@@ -64,6 +70,25 @@ BeforeAll {
         # Make sure two levels of objects are returned
         $script:result.reporter | Get-Member 'login' | Should -Not -BeNullOrEmpty
         $script:result.reporter.login | Should -Be 'admin'
+
+        if ($HasCustomFields)
+        {
+            foreach ($expectedFieldName in $HasCustomFields.Keys)
+            {
+                $expectedField = $script:result.customFields | Where-Object 'name' -eq $expectedFieldName
+                $actualField = $script:result | Get-YTIssueCustomField -Session $script:session -Field $expectedFieldName -Type $expectedField.'$type'
+                $actualField | Should -Not -BeNullOrEmpty
+                $actualField.name | Should -Be $expectedFieldName
+                if ($actualField.'$type' -eq 'SingleUserIssueCustomField')
+                {
+                    $actualField.value.login -eq 'admin'
+                }
+                else
+                {
+                    $actualField.value.name | Should -Be $HasCustomFields[$expectedFieldName]
+                }
+            }
+        }
     }
 }
 
@@ -77,16 +102,18 @@ Describe 'New-YTIssue' {
             Summary = 'First YTAutomation Issue'
             Description = 'This is the first issue created by the YouTrackAutomation module.'
         }
-        ThenIssue -Summary 'First YTAutomation Issue' `
-                  -Description 'This is the first issue created by the YouTrackAutomation module.'
+        ThenIssue -HasSummary 'First YTAutomation Issue' `
+                  -HasDescription 'This is the first issue created by the YouTrackAutomation module.'
     }
 
     It 'should allow issues with the same summary and description' {
         WhenCreatingIssue -WithArgs @{ Summary = 'same summary' ; Description = 'same description' }
-        ThenIssue -Summary 'same summary' -Description 'same description' -Project $script:project.shortName
+        ThenIssue -HasSummary 'same summary' -HasDescription 'same description'
         $initialIssueId = $script:result.idReadable
         WhenCreatingIssue -WithArgs @{ Summary = 'same summary' ; Description = 'same description' }
-        ThenIssue -Summary 'same summary' -Description 'same description' -Project $script:project.shortName
+        ThenIssue -HasSummary 'same summary' `
+                  -HasDescription 'same description' `
+                  -InProject $script:project.shortName
         $script:result.idReadable | Should -Not -Be $initialIssueId
     }
 
@@ -101,5 +128,24 @@ Describe 'New-YTIssue' {
         WhenCreatingIssue -WithArgs @{ Summary = 'Missing Parent'; Parent = 'fubarsnafu' } -ErrorAction SilentlyContinue
         $script:result | Should -BeNullOrEmpty
         $Global:Error | Should -Match 'not found|does not exist'
+    }
+
+    It 'sets a custom field' {
+        $customField = @{ name = 'Priority' ; value = @{ name = 'Major' } ; '$type' = 'SingleEnumIssueCustomField' }
+
+        WhenCreatingIssue -WithArgs @{ Summary = 'Adding custom fields.' ; CustomField = $customField }
+        ThenIssue -HasSummary 'Adding custom fields.' `
+                  -HasCustomFields @{ Priority = 'Major' }
+    }
+
+    It 'sets custom fields' {
+        $customFields = @(
+            @{ name = 'Priority' ; value = @{ name = 'Major' } ; '$type' = 'SingleEnumIssueCustomField' }
+            @{ name = 'Assignee' ;  value = @{ login = 'admin' } ; '$type' = 'SingleUserIssueCustomField' }
+        )
+
+        WhenCreatingIssue -WithArgs @{ Summary = 'Adding custom fields.' ; CustomField = $customFields }
+        ThenIssue -HasSummary 'Adding custom fields.' `
+                  -HasCustomFields @{ Priority = 'Major' ; Assignee = 'admin' }
     }
 }