+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * Function `Get-YTBundle` for getting a YouTrack bundle.
 * Pipe issue objects, issue IDs, or issue readable IDs to `Get-YTIssueCustomField`.
 * `Get-YTIssueCustomField` returns all object properties on custom field when given custom field type.
+* `New-YTIssue` sets custom fields on issue creation. Pass custom fields to new `CustomField` parameter.
 
 ### Changed